requirements.txt
streamlit
crewai
pysqlite3-binary
python-dotenv
crewai-tools
google-generativeai

